Improve property conversion functions
The inflate and deflate functions should be capable of handling
properties that have a primitive value and sub-properties at the same
time. When such a property is discovered, we store the property's
primitive value under its name prefixed with @ to make space for
sub-properties in the object tree. Both inflate and deflate this
conversion.